refactor: rename BeneficiaryCreate to BeneficiaryUpdate as it used only for update
