+++ app/src/lib/ui/index.ts
@@ -1,4 +1,4 @@
-export { default as ProBeneficiaryCreate } from './ProBeneficiaryCreate/index.svelte';
+export { default as ProBeneficiaryUpdate } from './ProBeneficiaryUpdate/index.svelte';
 export { default as ProBeneficiaryCard } from './ProBeneficiaryCard.svelte';
 export { default as ProBeneficiarySearchBar } from './ProBeneficiarySearchBar.svelte';
 export { default as LayerCDB } from './LayerCDB.svelte';
